backend/requirements.txt
django 
djangorestframework 
django-cors-headers 
psycopg2-binary 
python-dotenv
gunicorn
celery[redis]
redis
django-celery-beat
django-celery-results
requests
httpx
python-decouple
cryptography
django-filter
channels[daphne]
channels_redis
